Try the default drive first
GOG version vs. ergonomy's version has a three byte difference on this function which causes ergonomy's version to ask for the drive letter. 
I was not able to figure out what to change to get it to match the GOG version.

As I'm going to be adding other code that changes offsets anyways I'm fine changing it like this to behave like the GOG version.